Adding flaming asteroid
I thought about it for a moment and had to research the difference between asteroids, comets, and meteoroids. It turns out flaming asteroids don’t exist and the closet approximation is a meteoroid entering the earth’s atmosphere and burning up.
http://en.wikipedia.org/wiki/Meteor#Meteor

An important lesson for the next version of the game.