Able to send/receive data from the BLE device
The incoming data still needs parsing but the format is visible (1 byte as operator code (0x04) specifying accelerometer data, 4 bytes for X axis, 4 bytes for Y axis, and 4 bytes for Z axis)
There is an issue that if commands are sent to quickly to the device they will override each other. It may be necessary to implement a confirmation return packet that will assert the previously send packet was interpreted.