When the BLE device is turned off
it does not execute the disconnect delegate. Instead it prints an error message. I will need to rewrite this section so that it calls the delegate in any case.

Using an SKLabel and a date to count up in milliseconds as the Time based score. The score will reset when a meteorite contacts the ship. Upon contact the controller will buzz.