A bunch of improvements in connectivity
The boolean to determine whether a controller is fully connected is publicly available.
Executing the disconnect delegate regardless of error condition ensures the app will receive the message of a missing peripheral and remove itself as a delegate.

Able to connect, disconnect, and reconnect, have other controllers with different Advertisement IDs  running simultaneously and have only the assigned Advertisement ID connect.